PS-9148 feature: Decoupled threading and caching functionality in query_cache
https://perconadev.atlassian.net/browse/PS-9148

Threading-related functionality extracted from the 'query_cache' class into
a separate 'dictionary_flusher_thread' class.

This new class now accepts an instance of existing 'query_cache' class as
a parameter of its constructor in a form of shared pointer.

Changed the way how these two objects are now initialized / deinitialized in
'component.cpp' ('component_init()' / 'component_deinit()' functions).
