+++ components/masking_functions/include/masking_functions/dictionary_flusher_thread_fwd.hpp
@@ -13,17 +13,18 @@
    along with this program; if not, write to the Free Software Foundation,
    51 Franklin Street, Suite 500, Boston, MA 02110-1335 USA */
 
-#ifndef MASKING_FUNCTIONS_QUERY_CACHE_FWD_HPP
-#define MASKING_FUNCTIONS_QUERY_CACHE_FWD_HPP
+#ifndef MASKING_FUNCTIONS_DICTIONARY_FLUSHER_THREAD_FWD_HPP
+#define MASKING_FUNCTIONS_DICTIONARY_FLUSHER_THREAD_FWD_HPP
 
 #include <memory>
 
 namespace masking_functions {
 
-class query_cache;
+class dictionary_flusher_thread;
 
-using query_cache_ptr = std::shared_ptr<query_cache>;
+using dictionary_flusher_thread_ptr =
+    std::unique_ptr<dictionary_flusher_thread>;
 
 }  // namespace masking_functions
 
-#endif  // MASKING_FUNCTIONS_QUERY_CACHE_FWD_HPP
+#endif  // MASKING_FUNCTIONS_DICTIONARY_FLUSHER_THREAD_FWD_HPP
