+++ components/masking_functions/include/masking_functions/query_cache.hpp
@@ -18,18 +18,11 @@
 
 #include "masking_functions/query_cache_fwd.hpp"
 
-#include <atomic>
-#include <condition_variable>
-#include <memory>
 #include <mutex>
-#include <optional>
 #include <shared_mutex>
 #include <string>
 
-#include <mysql/components/services/psi_thread.h>
-
 #include "masking_functions/bookshelf_fwd.hpp"
-#include "masking_functions/dictionary_fwd.hpp"
 #include "masking_functions/query_builder_fwd.hpp"
 
 namespace masking_functions {
@@ -37,8 +30,7 @@ namespace masking_functions {
 class query_cache {
  public:
   // passing unique_ptr by value to transfer ownership
-  query_cache(query_builder_ptr query_builder,
-              std::uint64_t flusher_interval_seconds);
+  query_cache(query_builder_ptr query_builder);
   query_cache(const query_cache &other) = delete;
   query_cache(query_cache &&other) = delete;
   query_cache &operator=(const query_cache &other) = delete;
@@ -61,22 +53,6 @@ class query_cache {
 
   mutable bookshelf_ptr dict_cache_;
   mutable std::shared_mutex dict_cache_mutex_;
-
-  std::uint64_t flusher_interval_seconds_;
-  std::atomic<bool> is_flusher_stopped_;
-  std::mutex flusher_mutex_;
-  std::condition_variable flusher_condition_var_;
-
-  PSI_thread_key psi_flusher_thread_key_;
-  my_thread_handle flusher_thread_;
-  my_thread_attr_t flusher_thread_attr_;
-  std::unique_ptr<THD> flusher_thd_;
-
-  void init_thd() noexcept;
-  void release_thd() noexcept;
-  void dict_flusher() noexcept;
-
-  static void *run_dict_flusher(void *arg);
 
   bookshelf_ptr create_dict_cache_internal() const;
   using shared_lock_type = std::shared_lock<std::shared_mutex>;
