+++ components/masking_functions/src/masking_functions/query_cache.cpp
@@ -15,79 +15,25 @@
 
 #include "masking_functions/query_cache.hpp"
 
-#include <chrono>
 #include <string_view>
-
-#include <mysql/components/services/log_builtins.h>
-#include <mysql/psi/mysql_thread.h>
-#include <mysqld_error.h>
-#include <sql/debug_sync.h>
-#include <sql/sql_class.h>
 
 #include "masking_functions/bookshelf.hpp"
 #include "masking_functions/command_service_tuple.hpp"
 #include "masking_functions/primitive_singleton.hpp"
 #include "masking_functions/query_builder.hpp"
 #include "masking_functions/sql_context.hpp"
-#include "masking_functions/sys_vars.hpp"
 
-extern REQUIRES_SERVICE_PLACEHOLDER(log_builtins);
-
-namespace {
+namespace masking_functions {
 
 using global_command_services = masking_functions::primitive_singleton<
     masking_functions::command_service_tuple>;
 
-constexpr const char psi_category_name[]{"masking_functions"};
-constexpr const char flusher_thd_psi_name[]{"masking_functions_dict_flusher"};
-constexpr const char flusher_thd_psi_os_name[]{"mf_flusher"};
-
-}  // anonymous namespace
-
-namespace masking_functions {
-
-query_cache::query_cache(query_builder_ptr query_builder,
-                         std::uint64_t flusher_interval_seconds)
+query_cache::query_cache(query_builder_ptr query_builder)
     : dict_query_builder_{std::move(query_builder)},
       dict_cache_{},
-      dict_cache_mutex_{},
-      flusher_interval_seconds_{flusher_interval_seconds},
-      is_flusher_stopped_{true} {
-  // we do not initialize m_dict_cache with create_dict_cache_internal() here
-  // as this constructor is called from the component initialization method
-  // and any call to mysql_command_query service may mess up with current THD
+      dict_cache_mutex_{} {}
 
-  // the cache will be loaded during the first call to one of the dictionary
-  // functions or by the flusher thread
-  if (flusher_interval_seconds_ > 0) {
-    PSI_thread_info thread_info{&psi_flusher_thread_key_,
-                                flusher_thd_psi_name,
-                                flusher_thd_psi_os_name,
-                                PSI_FLAG_SINGLETON,
-                                0,
-                                PSI_DOCUMENT_ME};
-    mysql_thread_register(psi_category_name, &thread_info, 1);
-    my_thread_attr_init(&flusher_thread_attr_);
-
-    const auto res =
-        mysql_thread_create(psi_flusher_thread_key_, &flusher_thread_,
-                            &flusher_thread_attr_, run_dict_flusher, this);
-
-    if (res != 0) {
-      LogComponentErr(INFORMATION_LEVEL, ER_LOG_PRINTF_MSG,
-                      "Cannot initialize dictionary flusher");
-    } else {
-      is_flusher_stopped_ = false;
-    }
-  }
-}
-
-query_cache::~query_cache() {
-  if (!is_flusher_stopped_) {
-    is_flusher_stopped_ = true;
-    flusher_condition_var_.notify_one();
-  }
-}
+query_cache::~query_cache() = default;
 
 bool query_cache::contains(const std::string &dictionary_name,
                            const std::string &term) const {
@@ -166,66 +112,6 @@ void query_cache::reload_cache() {
   }
 
   dict_cache_ = std::move(local_dict_cache);
-}
-
-void query_cache::init_thd() noexcept {
-  auto *thd = new THD;
-  my_thread_init();
-  thd->set_new_thread_id();
-  thd->thread_stack = reinterpret_cast<char *>(&thd);
-  thd->store_globals();
-  flusher_thd_.reset(thd);
-}
-
-void query_cache::release_thd() noexcept { my_thread_end(); }
-
-void query_cache::dict_flusher() noexcept {
-#ifdef HAVE_PSI_THREAD_INTERFACE
-  {
-    struct PSI_thread *psi = flusher_thd_->get_psi();
-    PSI_THREAD_CALL(set_thread_id)(psi, flusher_thd_->thread_id());
-    PSI_THREAD_CALL(set_thread_THD)(psi, flusher_thd_.get());
-    PSI_THREAD_CALL(set_thread_command)(flusher_thd_->get_command());
-    PSI_THREAD_CALL(set_thread_info)
-    (STRING_WITH_LEN("Masking functions component cache flusher"));
-  }
-#endif
-
-  while (!is_flusher_stopped_) {
-    std::unique_lock lock{flusher_mutex_};
-    const auto wait_started_at = std::chrono::system_clock::now();
-    flusher_condition_var_.wait_for(
-        lock, std::chrono::seconds{flusher_interval_seconds_},
-        [this, wait_started_at] {
-          return std::chrono::duration_cast<std::chrono::seconds>(
-                     std::chrono::system_clock::now() - wait_started_at) >=
-                     std::chrono::seconds{flusher_interval_seconds_} ||
-                 is_flusher_stopped_.load();
-        });
-
-    if (!is_flusher_stopped_) {
-      {
-        unique_lock_type dict_cache_write_lock{dict_cache_mutex_};
-        auto local_dict_cache{create_dict_cache_internal()};
-        if (local_dict_cache) {
-          dict_cache_ = std::move(local_dict_cache);
-        }
-      }
-
-      DBUG_EXECUTE_IF("masking_functions_signal_on_cache_reload", {
-        const char act[] = "now SIGNAL masking_functions_cache_reload_done";
-        assert(!debug_sync_set_action(current_thd, STRING_WITH_LEN(act)));
-      };);
-    }
-  }
-}
-
-void *query_cache::run_dict_flusher(void *arg) {
-  auto *self = reinterpret_cast<masking_functions::query_cache *>(arg);
-  self->init_thd();
-  self->dict_flusher();
-  self->release_thd();
-  return nullptr;
 }
 
 bookshelf_ptr query_cache::create_dict_cache_internal() const {
